+++ NEWS.md
@@ -1,8 +1,9 @@
 # consort 1.2.3
 
--   Allow user configuration of arrow grpahical paraemters and padding.
+-   Allow user configuration of arrow graphical parameters and padding with `set_consort_defaults()`.
 -   Allow custom bullet characters
--   Imroved documentations. 
+-   Figure styles will also be applied to grViz plot.
+-   Improved documentation. 
 -   Special thanks to @Ramsas88
 
 # consort 1.2.2